scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
lxml==4.9.3
ics==0.7.2
pytz==2023.3 
icalendar==5.0.12
pdfminer.six==20231228